Added class factory to generate concrete HLS stream validator
Added, “analysis”, a boolean flag, used to decide which validator to
instantiate:
if true then hlsAnalysisStreamRegressionValidator object is created
(regression with analysis based on Dashboard’s python code
(live-monitor repository) .
If false, hlsStreamRegressionValidator (regression without analysis).
Increased the tests timeout to 24 hrs.
