+++ lib/Adapters/AdapterFactory.js
@@ -25,7 +25,6 @@ var exitProcess = function(code) {
 }
 
 module.exports = {
-    getAdapter: getAdapter,
-    exitProcess: exitProcess
+    getAdapter: getAdapter
 }
