+++ lib/Adapters/RegressionAdapter/RegressionEngine.js
@@ -13,17 +13,18 @@ var fs=require('fs');
 var networkClient = require('./../../NetworkClientFactory').getNetworkClient();
 var m3u8 = require('m3u8');
 var logger = require('../../../common/logger');
-var validator = require('./hlsStreamRegressionValidator');
+//var validator = require('./hlsStreamRegressionValidator');
 var ErrorUtils = require('../../utils/error-utils');
 var glob = require('glob');
 const qio = require('q-io/fs');
 const process = require('process');
-const regerssionConfig = require('./RegressionConfig');
 const genUtils = require('../../utils/general-utils');
-var weblogger = logger.getLogger("RegressionAdapter-Web");
+var validatorFactory = require('./StreamRegressionValidatorFactory');
+
 var regressionAdapterConfig = config.get('regressionAdapter');
 var mediaServer = config.get('mediaServer');
 var entriesInfo = config.get('regressionAdapter').entries;
+
 
 let logheader1 = function(literals, ...values) {
 
@@ -50,15 +51,6 @@ class EntryInfo {
         // the promise is resolved by min flavor when it started to perform the validation
         this.promise_per_flavor = {};
 
-/*
-        this.regression_status = {
-            running: 'running',
-            ending_failed: 'failed',
-            ending_succeeded: 'succeeded',
-            stopped: 'stopped'
-        };
-        this.status = regression_status.running;
-*/
     }
 }
 
@@ -112,7 +104,7 @@ RegressionEngine.prototype.initEntries = function() {
                 return that.initChunklistEdgeIndexes(entryConfig.entryId, that.entries[entryConfig.entryId].data_warehouse_path);
             })
             .then(function () {
-                that.validators[entryConfig.entryId] = new validator(entryConfig, that.entries[entryConfig.entryId].flavors, that.entries[entryConfig.entryId].last_chunklist_index);
+                that.validators[entryConfig.entryId] = validatorFactory.getValidator(entryConfig, that.entries[entryConfig.entryId].flavors, that.entries[entryConfig.entryId].last_chunklist_index);
             })
             .then(function() {
                 return that.validators[entryConfig.entryId].init();
@@ -258,40 +250,7 @@ RegressionEngine.prototype.exitRegressionWithError = function(entryId, flavorId,
     return deferred.promise;
 }
 
-RegressionEngine.prototype.validateSingleFlavor = function (entryId, flavorId, iteration) {
 
-    let that = this;
-    let deferred = Q.defer();
-    var flavor_header = logheader1 `${entryId}${flavorId}${iteration}`;
-    let url = util.format("http://localhost:8080/kLive/smil:%s_all.smil/%s/chunklist.m3u8", entryId, flavorId);
-
-    try {
-        that.logger.debug('%s getting chunklist, url=[%s]', flavor_header, url);
-
-        return networkClient.read({url: url, timeout: 10000})
-            .then(function (content) {
-                return that.validators[entryId].addChunklist(content, flavorId, iteration);
-            })
-            .then(function (m3u8) {
-                that.logger.info('%s ***** SUCCESSFULLY VALIDATED ***** url=[%s]', flavor_header, m3u8.toString(), url);
-                deferred.resolve(m3u8);
-            })
-            .catch(function (errors) {
-                return that.exitRegressionWithError(entryId, flavorId, iteration, errors, url);
-            })
-            .then(function() {
-                deferred.reject();
-            })
-            .catch(function() {
-                deferred.reject();
-            });
-    }
-    catch (e) {
-        that.logger.error('%s failed to get chunklist', flavor_header)
-    }
-
-    return deferred.promise;
-}
 
 RegressionEngine.prototype.validateAllFlavors = function (entryId, flavorId) {
 
@@ -324,7 +283,7 @@ RegressionEngine.prototype.validateAllFlavors = function (entryId, flavorId) {
         .then( function() {
             that.logger.debug('%s starting validation iteration, getting stream data from \'packager\'', log_header);
             let p = _.map(that.entries[entryId].flavors, function(next_flavor) {
-                return that.validateSingleFlavor(entryId, next_flavor, index);
+                return that.validators[entryId].validateSingleFlavor(entryId, next_flavor, index);
             });
 
             return Q.all(p)
@@ -419,7 +378,7 @@ RegressionEngine.prototype.resolveUrl = function(url, params) {
             params.validate = this.isMinFlavor(params.entryId, params.flavorId);
             var extended_filename = this.nextChunklist(params.entryId, params.flavorId);
             params.fullpath = this.entries[params.entryId].fullpath[params.flavorId] + '/' + extended_filename;
-            weblogger.info('[%s-%s] <<<< iteration %s >>>> next chunklist to get from DATA WAREHOUSE, fullpath=%s', params.entryId, params.flavorId, this.entries[params.entryId].chunklist_index[params.flavorId],  params.fullpath);
+            this.logger.info('[%s-%s] <<<< iteration %s >>>> next chunklist to get from DATA WAREHOUSE, fullpath=%s', params.entryId, params.flavorId, this.entries[params.entryId].chunklist_index[params.flavorId],  params.fullpath);
 
             if (this.entries[params.entryId].chunklist_index[params.flavorId] ===
                 this.entries[params.entryId].initial_chunklist_index) {
@@ -432,7 +391,7 @@ RegressionEngine.prototype.resolveUrl = function(url, params) {
                         deferred.resolve();
                     })
                     .catch(function (err) {
-                        weblogger.error('[%s-%s] ##### is this is a bug??? validateAllFlavors() returned reject with err %s', params.entryId, params.flavorId, err);
+                        that.logger.error('[%s-%s] ##### is this is a bug??? validateAllFlavors() returned reject with err %s', params.entryId, params.flavorId, err);
                         deferred.reject(err);
                     });
             }
@@ -450,7 +409,7 @@ RegressionEngine.prototype.resolveUrl = function(url, params) {
         // playlist (master manifest)
         else if (params.filename.indexOf('playlist_w111.m3u8') > -1) {
             params.request_type.playlist = true;
-            weblogger.debug('>>>>> GETTING PLAYLIST, OF ENTRY %s, FLAVOR %s <<<<< %s', params.entryId, params.flavorId, params.filename );
+            this.logger.debug('>>>>> GETTING PLAYLIST, OF ENTRY %s, FLAVOR %s <<<<< %s', params.entryId, params.flavorId, params.filename );
             this.resetChunklistCount(params.entryId, params.flavorId);
 
             deferred.resolve();
@@ -465,7 +424,7 @@ RegressionEngine.prototype.resolveUrl = function(url, params) {
         // exact number of chunklist files with same names are downloaded.
         // otherwise there should be a mechanism to skip chunklists that are absent for one or more flavors
     } catch (err) {
-        weblogger.error("error parsing url %s. Error: %s", url, ErrorUtils.error2string(err));
+        this.logger.error("error parsing url %s. Error: %s", url, ErrorUtils.error2string(err));
         deferred.reject();
     }
