+++ lib/Adapters/RegressionAdapter/hlsStreamRegressionValidator.js
@@ -14,6 +14,8 @@ const crypto = require('crypto');
 const v_utility = require('./validatorUtility');
 const qio = require('q-io/fs');
 const ErrorUtils = require('../../utils/error-utils');
+var StreamRegressionValidatorBase = require('./StreamRegressionValidatorBase')
+
 
 function checksum (str, algorithm, encoding) {
     return crypto
@@ -37,32 +39,32 @@ let logheader1 = function(literals, ...values) {
    all the chunklists and segments (ts files), fetched from the packager for single Live entry, during regression test
    period.
  */
-function HlsStreamRegressionValidator(config, flavors, last_chunklist_index) {
+class HlsStreamRegressionValidator extends StreamRegressionValidatorBase {
 
-    this.logger = logger.getLogger('hlsStreamRegressionValidator');
-    this.entryId = config.entryId;
-    this.flavors = flavors;
-    this.config = config.validator;
-    this.override = config.validator.override;
-    this.last_chunklist_index = last_chunklist_index;
-    // path to load/save regression test results
-    // the path is initialized in init()
-    this.fullpath = config.validator.path + '/' + config.entryPath;
-    if (config.entryPath.lastIndexOf('/') < this.fullpath.length - 1) {
-        this.fullpath += '/';
+    constructor (config, flavors, last_chunklist_index) {
+        super(config, flavors, last_chunklist_index);
+
+        this.logger = logger.getLogger('hlsStreamRegressionValidator');
+        this.override = config.validator.override;
+        this.last_chunklist_index = last_chunklist_index;
+        // path to load/save regression test results
+        // the path is initialized in init()
+        this.fullpath = config.validator.path + '/' + config.entryPath;
+        if (config.entryPath.lastIndexOf('/') < this.fullpath.length - 1) {
+            this.fullpath += '/';
+        }
+        // add root to path
+        this.db_filename = this.fullpath + this.entryId + '_regression_test_db.json';
+
+        // create initial json for regression results database
+        this.regression_results_db;
+        this.this_run_results = {};
     }
-    // add root to path
-    this.db_filename = this.fullpath + this.entryId + '_regression_test_db.json';
-
-    // create initial json for regression results database
-    this.regression_results_db;
-    this.this_run_results = {};
 
 }
 
 HlsStreamRegressionValidator.prototype.init = function() {
-
-
+    
     var that = this;
     return this.loadResultsDatabase()
         .then( function() {
@@ -180,6 +182,41 @@ HlsStreamRegressionValidator.prototype.addChunklist = function (chunklist, flavo
                 deferred.reject(data.errors);
             }
         });
+
+    return deferred.promise;
+}
+
+HlsStreamRegressionValidator.prototype.validateSingleFlavor = function (entryId, flavorId, iteration) {
+
+    let that = this;
+    let deferred = Q.defer();
+    var flavor_header = logheader1 `${entryId}${flavorId}${iteration}`;
+    let url = util.format("http://localhost:8080/kLive/smil:%s_all.smil/%s/chunklist.m3u8", entryId, flavorId);
+
+    try {
+        that.logger.debug('%s getting chunklist, url=[%s]', flavor_header, url);
+
+        return networkClient.read({url: url, timeout: 10000})
+            .then(function (content) {
+                return that.addChunklist(content, flavorId, iteration);
+            })
+            .then(function (m3u8) {
+                that.logger.info('%s ***** SUCCESSFULLY VALIDATED ***** url=[%s]', flavor_header, m3u8.toString(), url);
+                deferred.resolve(m3u8);
+            })
+            .catch(function (errors) {
+                return that.exitRegressionWithError(entryId, flavorId, iteration, errors, url);
+            })
+            .then(function() {
+                deferred.reject();
+            })
+            .catch(function() {
+                deferred.reject();
+            });
+    }
+    catch (e) {
+        that.logger.error('%s failed to get chunklist', flavor_header)
+    }
 
     return deferred.promise;
 }
