+++ lib/grunt/config/regression-tests-config.js
@@ -11,7 +11,7 @@ module.exports = function (grunt) {
             options: {
                 reporter: 'spec',
                 require: 'chai',
-                timeout: 43200000,
+                timeout: 86400000,
                 captureFile: '<%=reports_dir%>/regression-tests-report.xml'
             },
             src: ['<%=regression_tests%>']
@@ -20,7 +20,7 @@ module.exports = function (grunt) {
             options: {
                 reporter: 'xunit',
                 require: 'chai',
-                timeout: 43200000,
+                timeout: 86400000,
                 quiet : true,
                 captureFile: '<%=reports_dir%>/regression-tests-report.xml'
