+++ tests/regression/regression-spec.js
@@ -15,7 +15,7 @@ var ControllerWrapper =  require('./ControllerWrapper');
 describe('Regression test for HLS stream', function() {
 
     // Set an appropriate test timeout here to 12 hours: 12 * 3600 * 1000 = 43200000
-    this.timeout(43200000);
+    // this.timeout(43200000);
 
     it('should run regression and receive same result as ground truth', function(done) {
        var controller = new ControllerWrapper("regression test");
